Remove left-in class attribute defines from the merge; move Operation to nodes subfolder (since it's now a node).
