+++ earthmover/earthmover.py
@@ -241,10 +241,6 @@ class Earthmover:
                     node.execute()  # Sets self.data in each node.
                     node.post_execute()
 
-                # this doesn't seem necessary? row/col numbers are on the graph (after completion)
-                # if node.type=='transformation' and [cfg.get('debug', False) for cfg in node.config]:
-                #     node.num_rows = dask.compute(node.num_rows)[0]
-
 
     def generate(self, selector):
         """
